Add manual cut, bad det printout
TCutG and TVector are added for pre-defined cut. fun2 is defined over larger range and has better starting parameters. Note: bin with is still assumed to be 0.1 and does not work. Fix: Average slope function not summed. Back channels are not used, so summing is not used (and average slope is a misnomer). Arrays for counting bad detectors is added. Output front channel index is +8 instead of +4. Output range will be 8–11 instead of 4–7. File output loop moved to within DetNum loop.