Remove QQQ Method 2, Step 3
Remove Step 3 from Method 2 in QQQ relative cal. It appears that Step 3 was copied from the SX3 relative calibration folder. The contents of “Step 3” were nearly identical to Method 2 of SX3 Step 1. Print cut is added to Step 1 in SX3. Note, it appears that at some point the procedure for the SX3s was to calibrate up vs down last. This does not make sense… to the author.